+++ src/sumbasic/__init__.py
@@ -19,6 +19,9 @@
 #  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
 #  MA 02110-1301, USA.
 #  
-__version__ = "0.1.0a1";
+__version__ = "0.1.0a2";
 from .interpreter import BasicError, BasicInterpreter;
 from .program import BasicProgram;
+
+from .channels import ChannelError, ChannelManager;
+from .database import BasicDatabase;